refactor: layout 폴더 이름을 대문자로 변경
