+++ src/stories/CommonPageLayoutDecorator.tsx
@@ -1,6 +1,6 @@
 import { type Decorator } from '@storybook/react';
 
-import { LAYOUT_ROOT_ID } from '@/components/layout';
+import { LAYOUT_ROOT_ID } from '@/components/Layout';
 
 export const CommonPageLayoutDecorator: Decorator = Story => (
   <div id={LAYOUT_ROOT_ID} className='h-[844px] w-full shrink-0 shadow-xl'>
